Fix subevent data load
