+++ pages/events.js
@@ -4,14 +4,10 @@ import SubEvents from "../components/events/SubEvents";
 import { withRouter } from "next/router";
 import eventsData from "../data/global";
 import React, { Component } from "react";
-import SubEvent from "../components/events/SubEvent";
-import Mask from "../components/Mask";
 import AerodynamixContent from "../components/events/AerodynamixContent";
 import NavBar from "../components/Navbar";
 import MobileNav from "../components/MobileNav";
 import IconEvents from "../components/IconEvents";
-import axios from "axios";
-
 class Page extends Component {
   color = {
     cyberquest: "#1F3C68",
@@ -26,30 +22,30 @@ class Page extends Component {
     rasayans: "#212121",
     monopoly: "#212121"
   };
-  state = {
-    showModal: false,
-    subEvents: [],
-    subEventData: [],
-    color: "black"
-  };
-  fetchSubEvents = () => {
-    const { router } = this.props;
-    const eventName = router.query.name;
-    if (eventsData[eventName] && eventName !== "aerodynamix") {
-      axios
-        .get(`/static/data/${eventName}.json`)
-        .then(res => {
-          const subEvents = res.data.events;
-          this.setState({ subEvents });
-        })
-        .catch(function(err) {
-          console.log(err);
-        });
-    }
-  };
-  componentDidMount() {
-    this.fetchSubEvents();
-  }
+  // state = {
+  //   showModal: false,
+  //   //subEvents: [],
+  //   subEventData: [],
+  //   color: "black"
+  // };
+  // fetchSubEvents = () => {
+  //   const { router } = this.props;
+  //   const eventName = router.query.name;
+  //   if (eventsData[eventName] && eventName !== "aerodynamix") {
+  //     axios
+  //       .get(`/static/data/${eventName}.json`)
+  //       .then(res => {
+  //         const subEvents = res.data.events;
+  //         this.setState({ subEvents });
+  //       })
+  //       .catch(function(err) {
+  //         console.log(err);
+  //       });
+  //   }
+  // };
+  // componentDidMount() {
+  //   this.fetchSubEvents();
+  // }
 
   // static async getInitialProps({ query, req }) {
   //   let eventName = query.name;
@@ -61,21 +57,22 @@ class Page extends Component {
   //   }
   //   return { data: null };
   // }
-  showEventModal = (event, color) => {
-    const subEventData = this.state.subEvents.find(function(element) {
-      return element.name == event;
-    });
-    this.setState({
-      color,
-      showModal: true,
-      subEventData: [subEventData]
-    });
-  };
-  hideModal = () => {
-    this.setState({
-      showModal: false
-    });
-  };
+
+  // showEventModal = (event, color) => {
+  //   const subEventData = this.state.subEvents.find(function(element) {
+  //     return element.name == event;
+  //   });
+  //   this.setState({
+  //     color,
+  //     showModal: true,
+  //     subEventData: [subEventData]
+  //   });
+  // };
+  // hideModal = () => {
+  //   this.setState({
+  //     showModal: false
+  //   });
+  // };
 
   render() {
     const { router } = this.props;
@@ -108,19 +105,16 @@ class Page extends Component {
         <Meta color={"#212121"} />
         <NavBar path={router.pathname} color={"#212121"} />
         <MobileNav path={router.pathname} color={"#212121"} />
-        <EventBanner eventName={eventName} />
-        <SubEvents
-          showEventModal={this.showEventModal}
-          subEvents={this.state.subEvents}
-        />
+        <EventBanner eventName={eventName} router={this.props.router} />
+        <SubEvents router={this.props.router} />
         {/* <ToggleDisplay show={this.state.showModal}> */}
-        <SubEvent
+        {/* <SubEvent
           subEventData={this.state.subEventData}
           show={this.state.showModal}
           color={this.state.color}
           hideModal={this.hideModal}
         />
-        <Mask show={this.state.showModal} hideModal={this.hideModal} />
+        <Mask show={this.state.showModal} hideModal={this.hideModal} /> */}
       </div>
     );
   }
